+++ korlibs-image-core/src@linux/korlibs/image/core/CoreImage.linux.kt
@@ -9,6 +9,6 @@ object LinuxCoreImageFormatProvider : CoreImageFormatProvider {
     override suspend fun decode(data: ByteArray): CoreImage =
         StbiCoreImageFormatProvider.decode(data)
 
-    override suspend fun encode(image: CoreImage, format: CoreImageFormat, level: Float): ByteArray =
+    override suspend fun encode(image: CoreImage, format: CoreImageFormat, level: Double): ByteArray =
         StbiCoreImageFormatProvider.encode(image, format, level)
 }
